Start of some updates for the slideshow.
Also breaking out shortcakes into their own class to keep things
organized.
